refactor: 불필요한 파일 삭제 및 util/type 코드 분리
