--- app/QueryWrapper.tsx
@@ -1,16 +1,0 @@
-"use client";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-
-interface Props {
-  children: React.ReactNode;
-}
-
-const queryClient = new QueryClient();
-
-export const QueryWrapper = ({ children }: Props) => {
-  return (
-    <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
-  );
-};
-
-export default QueryWrapper;
