+++ types/category.ts
@@ -3,3 +3,9 @@ export interface Category {
   name: string;
   isFavorite: boolean;
 }
+
+export interface CategoryResponse {
+  id: number;
+  name: string;
+  isFavorite: boolean;
+}
