+++ types/comment.ts
@@ -8,3 +8,14 @@ export interface Comment {
   createdAt: string;
   isEdit: boolean;
 }
+
+export interface CommentResponse {
+  id: number;
+  member: {
+    id: number;
+    nickname: string;
+  };
+  content: string;
+  createdAt: string;
+  updatedAt: string;
+}
